pyproject.toml
[project]
name = "aon-obsidian-import"
version = "0.1.0"
description = "Import selected PF2e AoN entries as compact markdown"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.0",
  "httpx>=0.27.0",
  "lxml>=5.0.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.0",
  "typer>=0.12.0",
]

[project.scripts]
aon-import = "aon_import.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
